Implement rename functionality in chat server
The commit adds a rename feature that enables users to change their username in the chat server. First, it checks if the new username is already in use. If not, it locks the server map, removes the old username, adds the user back with the new username, then releases the lock. A confirmation message about the username upgrade is then sent to the user.