Add new Client.go, update Readme.md with project details
Implemented a new Client struct in client.go, establishing a connection to the server. Also added instructions and functionalities to the Readme.md file. This includes how to run the program, as well as details on all the available user actions such as renaming, broadcasting messages, and sending private messages.